Add a demo/exploratory perl_map_int_body class
This has shown a few things: that the body class for parallel_for is type-
dependent on the range type passed in (hence the type-specific name
perl_map_int_body), and that all these 'const' qualifiers will sure make
for a lot of deep copying.  Also, that callbacks are not necessarily re-
quired; method calls may be a tidier solution and alleviate the need for
$Storable::Deparse/Eval under the hood.  That would also require passing
the concurrent vector as an explicit argument to the perl_map_int_body
constructor.  A clunky API but probably a simple place to start.

This code compiles, and that's about the best you can say of it at the
moment :-)