Initial thread setup: pass through @INC transparently
Quite a lot for this to work.  First, we assign a unique sequence number to
the perl_tbb_init objects, guarded by a mutex.  Add a function to collect
the @INC from the master interpreter and set the C++ member.  In the inter-
preter pool grab function, we now set up the @INC before the first module
is loaded via load_module().  Additionally, we use the
@threads::tbb::worker array to store a flag that says that a unique
perl_tbb_init object has been initialized in that worker thread.  While
we're there, add a $threads::tbb::worker indicator, and a stub for loading
modules.