An initial implementation of the lazy deep copy function
"clone_other_sv" implements a form of the visitor pattern in Perl.  It
should be capable of walking memory structures "belonging" to another in-
terpreter, and producing a local version of them.  Almost all of the func-
tion is untested by the test suite; it needs to be tested with more complex
data structures etc.  But a useful starting point.