Rename threads::tbb::map_int_body to ::for_int_func
It's a type of tbb body function, which maps over an int range and array
and calls a static function.  While not as elegant as a code reference, we
were at least able to do away with all of the magic globals and take the
test case out of the module itself.  So, getting close to a part of a final
API.