Add a test case for the PVMG support (and put on the SYNOPSIS)
The new example is better, in that it is less restricted.  Make it take the
intro, and also add a script which actually shows it in action.  For this
degenerate test case, the "main" thread will process items about 18 times
the speed of worker threads.  Clearly, a more CPU-intensive example is
required :-).