Be sure to initialize new perl_concurrent_slot's with thingy = 0
You can really get used to C++'s habit of initializing everything for you,
but I was assuming that pointers automatically get set to 0 when created.
This doesn't appear to be the case: make it explicit (and to be paranoid,
also check sanity twice in the branch which found the bug).