Always clone the invocant to all threads, even the master one.
This is a single object so it should be fine.  The reason for this is that
the calling convention for object invocation is that you *are* allowed to
change state in the object.  But the invocant must not be altered as at any
time another thread can come along and clone it - if it is mid-update on a
non-concurrent operation (eg, a hash growing) then it is likely to crash.

It would break common sense too much to disallow modifications in the
invocant, so given that we are caching the cloned versions, we may as well
clone it for the master thread, too.