perl_concurrent_array: add another method to avoid double-free
Set the IV pointer to 0 to mark the fact it's been freed.  This is a quite
hack-ish way to avoid a double free; it means that the object was freed too
early.  We will need to be maintaining correct ref counts on all objects,
as well as updating them on clone, for a proper fix.