Work around strangeness with the tbb_interpreter_numbers global
I won't pretend to understand why, but for some reason, values going into
this concurrent_hash_map are sometimes not available later.  The find me-
thod returns false, even though according to the debug output, it has al-
ready been set (and the locks explicitly released) some time ago.  This is
a compromise; fall back to $threads::tbb::worker (initializing to 0 in the
master thread) if the global doesn't work, and for the other place where it
might not be found, give up and leak.