Push const-ness down deep into the clone_other_sv function
A good thing I did this, because I found various macros have side-effects,
such as SvIV, SvPV etc upgrading the things they are passed.  The const
marker makes all that illegal, so we must instead use SvIVX etc.  Not sure
why some variables are coming in as eg, SvTYPE(sv) == SVt_PVIV but not POK
(SvPVX == 0), but it matches up with an earlier oddness I noticed so it
seems reasonable to ignore the dual-variable nature when the SvPVX pointer
is null.