Free values in perl_concurrent_vector's
Here is another significant source of leaks; while we happily freed values
when they were replaced by a new value in slots, we also need to free them
when the vector is destroyed.  We can't put it on the slot, as these
objects are created and destroyed in many places (including calling free).