Also free values in for_int_method body function
This still isn't enough to get the test to pass; there is still a leak
somewhere, but it doesn't look to be the fault of the refcounter
module so mark as TODO.  It has at least protected from a double free
and allowed the objects to be passed, so this will have to do to begin
with.  Show it at least calling the original destructor in the case
where there was no transporting ever occurring.